+++ jstests/aggregation/sources/setWindowFields/linear_fill.js
@@ -11,6 +11,10 @@ load("jstests/aggregation/extras/window_function_helpers.js");
 load("jstests/aggregation/extras/utils.js");  // For arrayEq.
 load("jstests/libs/feature_flag_util.js");    // For isEnabled.
 
+if (!FeatureFlagUtil.isEnabled(db, "Fill")) {
+    jsTestLog("Skipping as featureFlagFill is not enabled");
+    return;
+}
 const coll = db.linear_fill;
 coll.drop();
