+++ jstests/aggregation/variables/search_meta.js
@@ -10,6 +10,12 @@
  */
 (function() {
 "use strict";
+const getSearchMetaParam = db.adminCommand({getParameter: 1, featureFlagSearchMeta: 1});
+const isSearchMetaEnabled = getSearchMetaParam.hasOwnProperty("featureFlagSearchMeta") &&
+    getSearchMetaParam.featureFlagSearchMeta.value;
+if (!isSearchMetaEnabled) {
+    return;
+}
 
 const coll = db.searchCollector;
 coll.drop();
@@ -40,7 +46,7 @@ const response = db.runCommand({
     cursor: {}
 });
 if (!response.ok) {
-    assert.commandFailedWithCode(response, [31082, 6047401] /* mongos or community */);
+    assert.commandFailedWithCode(response, [31082, 40324] /* community or mongos */);
 } else {
     assert.eq(response.cursor.firstBatch, []);
 }
