+++ jstests/cqf/type_bracket.js
@@ -2,6 +2,10 @@
 "use strict";
 
 load("jstests/libs/optimizer_utils.js");  // For checkCascadesOptimizerEnabled.
+if (!checkCascadesOptimizerEnabled(db)) {
+    jsTestLog("Skipping test because the optimizer is not enabled");
+    return;
+}
 
 const t = db.cqf_type_bracket;
 t.drop();
@@ -15,22 +19,22 @@ for (let i = 0; i < 100; i++) {
 {
     const res = t.explain("executionStats").aggregate([{$match: {a: {$lt: "2"}}}]);
     assert.eq(12, res.executionStats.nReturned);
-    assertValueOnPlanPath("PhysicalScan", res, "child.child.nodeType");
+    assert.eq("PhysicalScan", res.queryPlanner.winningPlan.optimizerPlan.child.child.nodeType);
 }
 {
     const res = t.explain("executionStats").aggregate([{$match: {a: {$gt: "95"}}}]);
     assert.eq(4, res.executionStats.nReturned);
-    assertValueOnPlanPath("PhysicalScan", res, "child.child.nodeType");
+    assert.eq("PhysicalScan", res.queryPlanner.winningPlan.optimizerPlan.child.child.nodeType);
 }
 {
     const res = t.explain("executionStats").aggregate([{$match: {a: {$lt: 2}}}]);
     assert.eq(2, res.executionStats.nReturned);
-    assertValueOnPlanPath("PhysicalScan", res, "child.child.nodeType");
+    assert.eq("PhysicalScan", res.queryPlanner.winningPlan.optimizerPlan.child.child.nodeType);
 }
 {
     const res = t.explain("executionStats").aggregate([{$match: {a: {$gt: 95}}}]);
     assert.eq(4, res.executionStats.nReturned);
-    assertValueOnPlanPath("PhysicalScan", res, "child.child.nodeType");
+    assert.eq("PhysicalScan", res.queryPlanner.winningPlan.optimizerPlan.child.child.nodeType);
 }
 
 assert.commandWorked(t.createIndex({a: 1}));
@@ -38,21 +42,21 @@ assert.commandWorked(t.createIndex({a: 1}));
 {
     const res = t.explain("executionStats").aggregate([{$match: {a: {$lt: "2"}}}]);
     assert.eq(12, res.executionStats.nReturned);
-    assertValueOnPlanPath("IndexScan", res, "child.leftChild.nodeType");
+    assert.eq("IndexScan", res.queryPlanner.winningPlan.optimizerPlan.child.leftChild.nodeType);
 }
 {
     const res = t.explain("executionStats").aggregate([{$match: {a: {$gt: "95"}}}]);
     assert.eq(4, res.executionStats.nReturned);
-    assertValueOnPlanPath("IndexScan", res, "child.leftChild.nodeType");
+    assert.eq("IndexScan", res.queryPlanner.winningPlan.optimizerPlan.child.leftChild.nodeType);
 }
 {
     const res = t.explain("executionStats").aggregate([{$match: {a: {$lt: 2}}}]);
     assert.eq(2, res.executionStats.nReturned);
-    assertValueOnPlanPath("IndexScan", res, "child.leftChild.nodeType");
+    assert.eq("IndexScan", res.queryPlanner.winningPlan.optimizerPlan.child.leftChild.nodeType);
 }
 {
     const res = t.explain("executionStats").aggregate([{$match: {a: {$gt: 95}}}]);
     assert.eq(4, res.executionStats.nReturned);
-    assertValueOnPlanPath("IndexScan", res, "child.leftChild.nodeType");
+    assert.eq("IndexScan", res.queryPlanner.winningPlan.optimizerPlan.child.leftChild.nodeType);
 }
 }());
